+++ esptest/jira/__init__.py
@@ -1,0 +1,21 @@
+"""Jira integration helpers."""
+
+from .attachments import (
+    AttachmentInfo,
+    download_all_attachments,
+    download_attachment,
+    list_attachments,
+    upload_attachments,
+)
+from .client import create_client, get_config, login_jira
+
+__all__ = [
+    'AttachmentInfo',
+    'create_client',
+    'download_all_attachments',
+    'download_attachment',
+    'get_config',
+    'list_attachments',
+    'login_jira',
+    'upload_attachments',
+]